Make dependencies not target-specific

+++ kapitan/dependency_manager/git.py
@@ -11,13 +11,14 @@ logger = logging.getLogger(__name__)
 
 class Git(Dependency):
     def __init__(self, source_uri, output_path, **kwargs):
-        super().__init__(source_uri, output_path)
+        super().__init__(source_uri, os.path.join(Dependency.dependencies_root_path, output_path))
         self.kwargs = kwargs
 
     def fetch(self):
         repo_cache_dirname = self._get_repo_cache_dirname()
-        if 'fresh_fetch' in self.kwargs:
-            self._clear_cache()
+        if self.kwargs.get('fetch_always', False):
+            self.clear_cache()
+
         repo = self._get_repo_cache_or_clone()
 
         if 'ref' in self.kwargs:
@@ -31,10 +32,10 @@ class Git(Dependency):
             if os.path.isdir(os.path.join(full_subdir)):
                 repo_cache_dirname = full_subdir
             else:
-                raise GitSubdirNotFoundError("subdir {} not found in repo: {}".format(sub_dir, self.source_uri))
+                raise GitSubdirNotFoundError("Dependency {} : subdir {} not found in repo".format(self.source_uri, sub_dir))
 
         copy_tree(repo_cache_dirname, self.output_path)
-        logger.info("copied dependency from {} to {}".format(repo_cache_dirname, self.output_path))
+        logger.info("Dependency {} : copied from {} to {}".format(self.source_uri,  repo_cache_dirname, self.output_path))
 
     def _get_repo_cache_dirname(self):
         repo_name = os.path.basename(self.source_uri)
@@ -42,17 +43,18 @@ class Git(Dependency):
         return repo_cache_dir
 
     def _get_repo_cache_or_clone(self):
-        repo_cache_dir = os.path.join(self._get_repo_cache_dirname())
+        repo_cache_dir = self._get_repo_cache_dirname()
         if os.path.isdir(repo_cache_dir):
             repo = Repo(repo_cache_dir)
-            logger.info('Repository cache loaded from {}'.format(repo_cache_dir))
+            logger.info('Dependency {} : cache loaded from {}'.format(self.source_uri, repo_cache_dir))
         else:
-            logger.info("Fetching dependency at {}".format(self.source_uri))
+            logger.info("Dependency {} : fetching now".format(self.source_uri))
             repo = Repo.clone_from(self.source_uri, repo_cache_dir)
-            logger.info('Repository successfully fetched from {}'.format(self.source_uri))
+            logger.info('Dependency {} : successfully fetched'.format(self.source_uri))
         return repo
 
-    def _clear_cache(self):
+    def clear_cache(self):
         repo_cache_dir = self._get_repo_cache_dirname()
-        remove_tree(repo_cache_dir)
-        logger.info("Removed cache in {}".format(repo_cache_dir))
+        if os.path.isdir(repo_cache_dir):
+            remove_tree(repo_cache_dir)
+            logger.info("Dependency {} : removed cache from {}".format(self.source_uri, repo_cache_dir))
